+++ src/ModLayers/MenuLayer/ModifyFiles.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include "ModifyHooks/addbuttons.hpp"
+#include "ModifyHooks/jokesettings.hpp"
+#include "ModifyHooks/MenuManipulation.hpp"
+#include "ModifyHooks/Shortcuts.hpp"
+#include "ModifyHooks/hideshow.hpp"