+++ src/ModLayers/MenuLayer/ModifyHooks/addbuttons.hpp
@@ -1,0 +1,27 @@
+#include <Geode/Geode.hpp>
+#include <UIBuilder.hpp>
+using namespace geode::prelude;
+static void addbuttons(auto layer) {
+if (Mod::get()->getSettingValue<bool>("NoNewGroundsButton")) {
+                if (auto newgrounds = layer->getChildByID("bottom-menu")->getChildByIDRecursive("newgrounds-button"))
+                { 
+                    newgrounds->removeFromParent();
+                };
+              }
+              if (Mod::get()->getSettingValue<bool>("MoreGamesBottomMenu")) {
+         
+            auto btn = CircleButtonSprite::createWithSprite(
+                "epic_text.png"_spr,
+                1.15f,
+                CircleBaseColor::Green,
+                CircleBaseSize::MediumAlt
+            );
+            
+		        auto btnee = CCMenuItemSpriteExtra::create(btn, layer, menu_selector(MenuLayer::onMoreGames));
+
+            layer->getChildByID("bottom-menu")->addChild(btnee);
+              };
+
+     layer->getChildByID("bottom-menu")->updateLayout();
+return;
+}