+++ changelog.md
@@ -1,4 +1,6 @@
 # ChangeLog
+### BetterMenu-beta.12
+* Fix Ui for prism menu and other buttons
 ### BetterMenu-beta.11
 * small screen bug fix for certain android screen sizes
 * icon profile works