+++ changelog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 ### BetterMenu-beta.12
 * Fix Ui for prism menu and other buttons
+* Fixed ui not being centered
 ### BetterMenu-beta.11
 * small screen bug fix for certain android screen sizes
 * icon profile works