+++ changelog.md
@@ -1,4 +1,8 @@
 # ChangeLog
+### BetterMenu V2.03
+* Cleaner Code
+* bug fixes
+* macos fully suports tweens
 ### BetterMenu V2.02
 * New games button setting
 * Newgrounds button remover setting